Hateoas

+++ dao/src/main/java/com/epam/mjc/dao/GiftCertificateDao.java
@@ -3,7 +3,6 @@ package com.epam.mjc.dao;
 import com.epam.mjc.dao.entity.GiftCertificateEntity;
 import com.epam.mjc.dao.entity.SearchParams;
 
-import java.math.BigInteger;
 import java.util.List;
 
 public interface GiftCertificateDao {
@@ -17,5 +16,5 @@ public interface GiftCertificateDao {
 
     Long create(GiftCertificateEntity item);
 
-    BigInteger countCertificates(SearchParams searchParams);
+    int countCertificates(SearchParams searchParams);
 }
