+++ service/src/main/java/com/epam/mjc/service/TagService.java
@@ -2,7 +2,6 @@ package com.epam.mjc.service;
 
 import com.epam.mjc.dao.dto.TagDto;
 
-import java.math.BigInteger;
 import java.util.List;
 
 public interface TagService {
@@ -11,5 +10,5 @@ public interface TagService {
     TagDto createTag(TagDto tagDto);
     String deleteTagById(Long id);
     TagDto getMostPopularAndExpensiveTag();
-    BigInteger countTags();
+    int countTags();
 }
