+++ service/src/main/java/com/epam/mjc/service/impl/GiftCertificateServiceImpl.java
@@ -18,7 +18,6 @@ import org.springframework.transaction.annotation.EnableTransactionManagement;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.math.BigDecimal;
-import java.math.BigInteger;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -97,7 +96,7 @@ public class GiftCertificateServiceImpl implements GiftCertificateService {
     }
 
     @Override
-    public BigInteger countOrders(SearchParams params) {
+    public int countOrders(SearchParams params) {
         return certificateDao.countCertificates(params);
     }
